requirements.txt
# Polymarket 体育赛事实时波动套利 — 依赖
# 目的：CLOB 交易、WebSocket、配置与测试

py-clob-client>=0.18.0
websocket-client>=1.6.0
requests>=2.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
